Fixed a bug in totalCost(), changed type to Bigdecimal; more appropriate for financial calculations

+++ src/com/main/Order.java
@@ -1,5 +1,7 @@
 package com.main;
 
+import java.math.BigDecimal;
+
 /**
  * Abstract class for specialised order types
  * to inherit from i.e. 'Basket' or 'Checkout
@@ -9,6 +11,6 @@ public abstract class Order {
     /**
      * Enforces subclasses have their own totalCost method
      */
-    public abstract double totalCost();
+    public abstract BigDecimal totalCost();
 
 }
